Rework Getting Started/projects page
