+++ .astro/settings.json
@@ -1,5 +1,8 @@
 {
 	"devToolbar": {
 		"enabled": false
+	},
+	"_variables": {
+		"lastUpdateCheck": 1727778140156
 	}
 }
